\name{mkHelpDir}
\alias{mkHelpDir}
\title{Collect HTML Documentation}
\description{
  \code{mkHelpDir} generates \code{HTML} documentation shown by \code{\link{help.start}}
  and puts it into a specified directory.
}
\usage{
mkHelpDir(path="~/R")
}
\arguments{
\item{path}{a character string; specifies the directory
where to put the documentation. By default the \code{"R"}
subdirectory in the user's home directory.}
}
\keyword{misc}